Refactoring code to fit more to SOLID principles.
