+++ src/Components/Graph Components/Vertex/Vertex.js
